+++ src/crypto/sha1.cpp
@@ -32,7 +32,7 @@ sha1::encoder::encoder() {
   reset();
 }
 
-sha1 sha1::hash( const char* d, uint32_t dlen ) {
+sha1 sha1::hash( const char* d, size_t dlen ) {
   encoder e;
   e.write(d,dlen);
   return e.result();
@@ -41,7 +41,7 @@ sha1 sha1::hash( const string& s ) {
   return hash( s.c_str(), s.size() );
 }
 
-void sha1::encoder::write( const char* d, uint32_t dlen ) {
+void sha1::encoder::write( const char* d, size_t dlen ) {
   SHA1_Update( &my->ctx, d, dlen); 
 }
 sha1 sha1::encoder::result() {
